examples/google-adk-a2a/pyproject.toml
[project]
name = "travel-agents"
version = "0.1.0"
description = "Weather and Trip Planner agents using A2A"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "a2a-sdk==0.3.3",
    "fastmcp==2.11.3",
    "google-adk==1.13.0",
    "google-genai>=1.17.0",
    "python-dotenv>=1.1.0",
    "httpx>=0.27.0",
    "uvicorn>=0.30.0",
    "bmasterai",
]
